+++ netpyne/plotting/plotTimeSeries.py
@@ -10,7 +10,6 @@ from .plotter import ImagePlotter
 from .plotter import MetaFigure
 import numpy as np
 
-
 @exception
 def plotLFPTimeSeries(
     LFPData=None,
@@ -19,7 +18,6 @@ def plotLFPTimeSeries(
     electrodes=['avg', 'all'],
     pop=None,
     separation=1.0,
-    logy=False,
     normSignal=False,
     filtFreq=False,
     filtOrder=3,
@@ -33,13 +31,6 @@ def plotLFPTimeSeries(
     **kwargs
 ):
     """Function to produce a line plot of LFP electrode signals
-
-    NetPyNE Options
-    ---------------
-    sim : NetPyNE sim object
-        The *sim object* from which to get data.
-
-        *Default:* ``None`` uses the current NetPyNE sim object
 
     Parameters
     ----------
@@ -74,11 +65,6 @@ def plotLFPTimeSeries(
         Use to increase or decrease distance between signals on the plot.
 
         *Default:* ``1.0``
-
-    logy : bool
-        Whether to use a log axis.
-
-        *Default:* ``False``
 
     normSignal : bool
         Whether to normalize the data.
@@ -178,8 +164,7 @@ def plotLFPTimeSeries(
         By default, returns the *figure*.  If ``returnPlotter`` is ``True``, instead returns the NetPyNE MetaFig.
 
     """
-
-    # If there is no input data, get the data from the NetPyNE sim object
+        # If there is no input data, get the data from the NetPyNE sim object
     if LFPData is None:
         if 'sim' not in kwargs:
             from .. import sim
@@ -192,7 +177,6 @@ def plotLFPTimeSeries(
             electrodes=electrodes,
             pop=pop,
             LFPData=LFPData,
-            logy=logy,
             normSignal=normSignal,
             filtFreq=filtFreq,
             filtOrder=filtOrder,
@@ -202,17 +186,321 @@ def plotLFPTimeSeries(
 
     print('Plotting LFP time series...')
 
-    # If input is a dictionary, pull the data out of it
-    if type(LFPData) == dict:
+    return plotTimeSeries(
+        data=LFPData,
+        label="LFPTimeSeries",
+        title='LFP Time Series Plot',
+        yLabel='LFP Signal (uV)',
+        axis=axis,
+        pop=pop,
+        separation=separation,
+        orderInverse=orderInverse,
+        overlay=overlay,
+        scalebar=scalebar,
+        legend=legend,
+        colorList=colorList,
+        returnPlotter=returnPlotter,
+        **kwargs)
 
-        t = LFPData['t']
-        lfps = LFPData['electrodes']['lfps']
-        names = LFPData['electrodes']['names']
-        locs = LFPData['electrodes']['locs']
-        colors = LFPData.get('colors')
-        linewidths = LFPData.get('linewidths')
-        alphas = LFPData.get('alphas')
-        axisArgs = LFPData.get('axisArgs')
+
+
+@exception
+def plotCSDTimeSeries(
+    CSDData=None,
+    axis=None,
+    timeRange=None,
+    electrodes=['avg', 'all'],
+    pop=None,
+    separation=1.0,
+    orderInverse=True,
+    overlay=False,
+    scalebar=True,
+    legend=True,
+    colorList=None,
+    returnPlotter=False,
+    **kwargs
+):
+    """"
+
+    Parameters
+    ----------
+    CSDData : dict, str
+        The data necessary to plot the CSD signals. 
+
+        *Default:* ``None`` uses ``analysis.prepareCSD`` to produce ``CSDData`` using the current NetPyNE sim object.
+
+        If a *str* it must represent a file path to previously saved data.
+
+    axis : matplotlib axis
+        The axis to plot into, allowing overlaying of plots.
+
+        *Default:* ``None`` produces a new figure and axis.
+
+    timeRange : list
+        Time range to include in the raster: ``[min, max]``.
+
+        *Default:* ``None`` uses the entire simulation
+
+    electrodes : list
+        A *list* of the electrodes to plot from.
+
+        *Default:* ``['avg', 'all']`` plots each electrode as well as their average
+
+    pop : str
+        A population name to calculate signals from.
+
+        *Default:* ``None`` uses all populations.
+
+    separation : float
+        Use to increase or decrease distance between signals on the plot.
+
+        *Default:* ``1.0``
+
+    orderInverse : bool
+        Whether to invert the order of plotting.
+
+        *Default:* ``True``
+
+    overlay : bool
+        Option to label signals with a color-matched overlay.
+
+        *Default:* ``False``
+
+    scalebar : bool
+        Whether to to add a scalebar to the plot.
+
+        *Default:* ``True``
+
+    legend : bool
+        Whether or not to add a legend to the plot.
+
+        *Default:* ``True`` adds a legend.
+
+    colorList : list
+        A *list* of colors to draw from when plotting.
+
+        *Default:* ``None`` uses the default NetPyNE colorList.
+
+    returnPlotter : bool
+        Whether to return the figure or the NetPyNE MetaFig object.
+
+        *Default:* ``False`` returns the figure.
+
+
+    Plot Options
+    ------------
+    showFig : bool
+        Whether to show the figure.
+
+        *Default:* ``False``
+
+    saveFig : bool
+        Whether to save the figure.
+
+        *Default:* ``False``
+
+    overwrite : bool
+        whether to overwrite existing figure files.
+
+        *Default:* ``True`` overwrites the figure file
+
+        *Options:* ``False`` adds a number to the file name to prevent overwriting
+
+    legendKwargs : dict
+        a *dict* containing any or all legend kwargs.  These include ``'title'``, ``'loc'``, ``'fontsize'``, ``'bbox_to_anchor'``, ``'borderaxespad'``, and ``'handlelength'``.
+
+    rcParams : dict
+        a *dict* containing any or all matplotlib rcParams.  To see all options, execute ``import matplotlib; print(matplotlib.rcParams)`` in Python.  Any options in this *dict* will be used for this current figure and then returned to their prior settings.
+
+    title : str
+        the axis title
+
+    xlabel : str
+        label for x-axis
+
+    ylabel : str
+        label for y-axis
+
+    linewidth : int
+        line width
+
+    alpha : float
+        line opacity (0-1)
+
+
+    Returns
+    -------
+    LFPTimeSeriesPlot : *matplotlib figure*
+        By default, returns the *figure*.  If ``returnPlotter`` is ``True``, instead returns the NetPyNE MetaFig.
+
+    """
+
+    """Function to produce a line plot of CSD electrode signals"""
+
+    # If there is no input data, get the data from the NetPyNE sim object
+    if CSDData is None:
+        if 'sim' not in kwargs:
+            from .. import sim
+        else:
+            sim = kwargs['sim']
+
+        CSDData = sim.analysis.prepareCSD(
+            sim=sim,
+            timeRange=timeRange, 
+            electrodes=electrodes,
+            pop=pop,
+            getAllData=False,
+            **kwargs)
+
+    print('Plotting CSD time series...')
+
+    from .plotTimeSeries import plotTimeSeries
+    return plotTimeSeries(
+        data=CSDData,
+        label='CSDTimeSeries',
+        title='CSD Time Series Plot',
+        yLabel=r'CSD Signal (mV/(mm$^2$))',
+        axis=axis,
+        pop=pop,
+        separation=separation,
+        orderInverse=orderInverse,
+        overlay=overlay,
+        scalebar=scalebar,
+        legend=legend,
+        colorList=colorList,
+        returnPlotter=returnPlotter,
+        **kwargs)
+
+
+@exception
+def plotTimeSeries(
+    data,
+    label='TimeSeries',
+    title='Time Series Plot',
+    yLabel=None,
+    axis=None,
+    pop=None,
+    separation=1.0,
+    orderInverse=True,
+    overlay=False,
+    scalebar=True,
+    legend=True,
+    colorList=None,
+    returnPlotter=False,
+    **kwargs
+):
+    """Function to produce a line plot of electrode signals (currently, LFP or CSD)
+
+    Parameters
+    ----------
+    data : dict, str
+        The data necessary to plot the signals.
+
+    axis : matplotlib axis
+        The axis to plot into, allowing overlaying of plots.
+
+        *Default:* ``None`` produces a new figure and axis.
+
+    pop : str
+        A population name to calculate signals from.
+
+        *Default:* ``None`` uses all populations.
+
+    separation : float
+        Use to increase or decrease distance between signals on the plot.
+
+        *Default:* ``1.0``
+
+    orderInverse : bool
+        Whether to invert the order of plotting.
+
+        *Default:* ``True``
+
+    overlay : bool
+        Option to label signals with a color-matched overlay.
+
+        *Default:* ``False``
+
+    scalebar : bool
+        Whether to to add a scalebar to the plot.
+
+        *Default:* ``True``
+
+    legend : bool
+        Whether or not to add a legend to the plot.
+
+        *Default:* ``True`` adds a legend.
+
+    colorList : list
+        A *list* of colors to draw from when plotting.
+
+        *Default:* ``None`` uses the default NetPyNE colorList.
+
+    returnPlotter : bool
+        Whether to return the figure or the NetPyNE MetaFig object.
+
+        *Default:* ``False`` returns the figure.
+
+
+    Plot Options
+    ------------
+    showFig : bool
+        Whether to show the figure.
+
+        *Default:* ``False``
+
+    saveFig : bool
+        Whether to save the figure.
+
+        *Default:* ``False``
+
+    overwrite : bool
+        whether to overwrite existing figure files.
+
+        *Default:* ``True`` overwrites the figure file
+
+        *Options:* ``False`` adds a number to the file name to prevent overwriting
+
+    legendKwargs : dict
+        a *dict* containing any or all legend kwargs.  These include ``'title'``, ``'loc'``, ``'fontsize'``, ``'bbox_to_anchor'``, ``'borderaxespad'``, and ``'handlelength'``.
+
+    rcParams : dict
+        a *dict* containing any or all matplotlib rcParams.  To see all options, execute ``import matplotlib; print(matplotlib.rcParams)`` in Python.  Any options in this *dict* will be used for this current figure and then returned to their prior settings.
+
+    title : str
+        the axis title
+
+    xlabel : str
+        label for x-axis
+
+    ylabel : str
+        label for y-axis
+
+    linewidth : int
+        line width
+
+    alpha : float
+        line opacity (0-1)
+
+
+    Returns
+    -------
+    time series plot : *matplotlib figure*
+        By default, returns the *figure*.  If ``returnPlotter`` is ``True``, instead returns the NetPyNE MetaFig.
+
+    """
+
+    # If input is a dictionary, pull the data out of it
+    if type(data) == dict:
+
+        t = data['t']
+        lfps = data['electrodes']['data']
+        names = data['electrodes']['names']
+        locs = data['electrodes']['locs']
+        colors = data.get('colors')
+        linewidths = data.get('linewidths')
+        alphas = data.get('alphas')
+        axisArgs = data.get('axisArgs')
 
     # Set up colors, linewidths, and alphas for the plots
     if not colors:
@@ -229,12 +517,12 @@ def plotLFPTimeSeries(
     # Create a dictionary to hold axis inputs
     if not axisArgs:
         axisArgs = {}
-        title = 'LFP Time Series Plot'
         if pop:
             title += ' - Population: ' + pop
         axisArgs['title'] = title
         axisArgs['xlabel'] = 'Time (ms)'
-        axisArgs['ylabel'] = 'LFP Signal (uV)'
+        if yLabel:
+            axisArgs['ylabel'] = yLabel
 
     # Link colors to traces, make avg plot black, add separation to traces
     plotColors = []
@@ -274,7 +562,7 @@ def plotLFPTimeSeries(
             kwargs.pop(kwarg)
 
     # create Plotter object
-    linesPlotter = LinesPlotter(data=linesData, kind='LFPTimeSeries', axis=axis, **axisArgs, **kwargs)
+    linesPlotter = LinesPlotter(data=linesData, kind=label, axis=axis, **axisArgs, **kwargs)
     metaFig = linesPlotter.metafig
 
     # Set up the default legend settings
@@ -338,10 +626,10 @@ def plotLFPTimeSeries(
             )
 
     # Generate the figure
-    LFPTimeSeriesPlot = linesPlotter.plot(**axisArgs, **kwargs)
+    timeSeriesPlot = linesPlotter.plot(**axisArgs, **kwargs)
 
     # Default is to return the figure, but you can also return the plotter
     if returnPlotter:
         return metaFig
     else:
-        return LFPTimeSeriesPlot
+        return timeSeriesPlot